requirements.txt
annotated-types==0.7.0
anyio==4.3.0
appnope==0.1.3
asn1crypto==1.5.1
astor==0.8.1
asttokens==2.4.1
boto3==1.34.144
botocore==1.34.144
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
croniter==2.0.5
cryptography==42.0.8
dateparser==1.2.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
duckdb==0.10.2
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.2
filelock==3.15.4
freezegun==1.5.1
h11==0.14.0
httptools==0.6.1
hyperscript==0.0.3
identify==2.5.36
idna==3.7
importlib_metadata==8.0.0
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.2
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab_widgets==3.0.10
keyring==24.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.3.0
nest-asyncio==1.5.8
nodeenv==1.9.1
numpy==1.26.4
packaging==23.2
pandas==2.2.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
pre-commit==3.7.1
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyGithub==2.3.0
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2024.5.15
requests==2.32.2
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.11.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
sqlglot==24.1.3
sqlglotrs==0.2.5
sqlmesh==0.101.0
sse-starlette==2.1.0
stack-data==0.6.3
starlette==0.37.2
tabula==1.0.5
tabula-py==2.9.3
tomlkit==0.12.5
tornado==6.4
traitlets==5.14.0
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.22.0
uvloop==0.19.0
virtualenv==20.26.3
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
widgetsnbextension==4.0.10
wrapt==1.16.0
zipp==3.19.2
